board: rockchip: Add Pine64 PineNote
The PineNote is an e-ink tablet by PINE64. It is built around a Rockchip
RK3566 SoC and has the following specifications:
- 4GB LPDDR4
- 128GB eMMC
- 10.3" 1404x1872 e-ink display

This board is very similar to other RK356x-based boards, with the
exception that its download-mode "key" (actually a Hall sensor) is on
ADC channel 0 instead of 1.

As such, this patch is heavily based on the evb_3568 board
implementation with the addition of a simple board setup code, meant to
account for the above difference.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
Signed-off-by: Jonas Karlman <jonas@kwiboo.se>